# API_M3_V5/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "API_M3",
    "description": "XTREME API",
    "version": "v5"
  },
  "paths": {
...
    "/Staff/StaffByChar": {
      "get": {
        "tags": [
          "Staff"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "type",
            "type": "string"
          },
          {
            "in": "query",
            "name": "value",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/Staff/NewStaff": {
      "post": {
        "tags": [
          "Staff"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Staff_aux"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/Staff/UpStaff": {
      "put": {
        "tags": [
          "Staff"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Staff_aux"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
    "/Supplier/SupplierByChar": {
      "get": {
        "tags": [
          "Supplier"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "type",
            "type": "string"
          },
          {
            "in": "query",
            "name": "value",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/Supplier/NewSupplier": {
      "post": {
        "tags": [
          "Supplier"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Supplier_aux"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Staff_aux": {
      "type": "object",
      "properties": {
        "employeeId": {
          "format": "int32",
          "type": "integer"
        },
        "username": {
          "type": "string"
        },
        "jobTitle": {
          "type": "string"
        },
        "personId": {
          "format": "int32",
          "type": "integer"
        },
        "entityId": {
          "format": "int32",
          "type": "integer"
        },
        "firstName": {
          "type": "string"
        },
        "surname": {
          "type": "string"
        },
        "nif": {
          "format": "int32",
          "type": "integer"
        },
        "email": {
          "type": "string"
        },
        "phone": {
          "format": "int32",
          "type": "integer"
        },
        "addressline": {
          "type": "string"
        },
        "zipcode": {
          "format": "int32",
          "type": "integer"
        },
        "locationId": {
          "format": "int32",
          "type": "integer"
        },
        "city": {
          "type": "string"
        },
        "district": {
          "type": "string"
        },
        "country": {
          "type": "string"
        }
      }
    },
    "Supplier_aux": {
      "type": "object",
      "properties": {
        "supplierId": {
          "format": "int32",
          "type": "integer"
        },
        "deliveryAverage": {
          "format": "double",
          "type": "number"
        },
        "businessId": {
          "format": "int32",
          "type": "integer"
        },
        "entityId": {
          "format": "int32",
          "type": "integer"
        },
        "businessType": {
          "type": "string"
        },
        "businessName": {
          "type": "string"
        },
        "nif": {
          "format": "int32",
          "type": "integer"
        },
        "addressline": {
          "type": "string"
        },
        "zipcode": {
          "type": "string"
        },
        "locationId": {
          "format": "int32",
          "type": "integer"
        },
        "city": {
          "type": "string"
        },
        "district": {
          "type": "string"
        },
        "country": {
          "type": "string"
        }
      }
    }
  }
}